dbsp_adapters: Fix and improve division of CSV into records.
Most of the time, one line is one record in CSV format, but it's possible
to span lines by enclosing new-lines in double quotes.  The adapter only
half-heartedly supported this because it allowed for records to span lines
only if they did not span fragments.  This commit fixes that.

This commit also fixes odd apparent limitations on the number of bytes and
fields per CSV record, since `csv_core::read_record` needed a buffer passed
in for these and we supplied 1024 and 128 of each.  Maybe the code was
written so this wasn't a limitation after all but it also avoids an extra
full copy of the data and 1024 + 128 * 8 = 2 kB of allocation and zeroing
per input fragment, which seems worthwhile.

Signed-off-by: Ben Pfaff <blp@feldera.com>