dbsp_adapters: Make the controller responsible for queuing batches.
Until now, the controller has only asked the transports to flush batches to
the circuit.  This commit instead makes the transports hand it batches as
they assemble them, and then the controller itself flushes them to the
circuit.

This wasn't possible before Nexmark shifted from three tables to one,
because I didn't see a clean way to make the controller flush the three
tables in synchronization.

Signed-off-by: Ben Pfaff <blp@feldera.com>