dbsp_adapters: Make `Parser` trait indicate record boundaries.
For fault tolerance, Feldera needs to be able to replay input.  To do
that, it also needs to be able to figure out record boundaries, so that
it can replay complete records instead of partial fragments.  Until now,
the `Parser` trait hasn't allowed this, because its support for fragments
is a black box that doesn't indicate record boundaries.

This commit refactors `Parser` so that, instead of supporting fragments,
it instead provides a way to split incoming data into records.  Then, for
fault tolerance, Feldera will work in two phases: first, split data into
records, then parse those records in full.

Signed-off-by: Ben Pfaff <blp@feldera.com>