dbsp_adapters: Change `self: &Arc<Self>` to just `&self` in some places.
The `&Arc<Self>` form is only needed in a case where the `Arc` itself is
important, for example if `self.clone()` or `Arc::downgrade(self)` is
going to be called.  Otherwise, I think that `&self` is simpler and
clearer.  The two forms are interchangeable for calls to methods (except
that the `Arc` form can only be called through an `Arc`).

Signed-off-by: Ben Pfaff <blp@feldera.com>